microservices/auth-service/requirements.txt
<<<<<<< HEAD
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic-settings==2.3.0
sqlalchemy[asyncio]==2.0.31
asyncpg==0.29.0
redis==5.0.4
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
twilio==9.2.3
tenacity==9.0.0
prometheus-fastapi-instrumentator==6.1.0
structlog==24.1.0
loguru==0.7.2
=======
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
twilio==8.12.0
httpx==0.25.2
python-dotenv==1.0.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-redis==0.42b0
opentelemetry-instrumentation-psycopg2==0.42b0
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk[fastapi]==1.39.1
pytest==7.4.3
pytest-asyncio==0.21.1
faker==20.1.0
bcrypt==4.2.0
PyJWT==2.8.0
>>>>>>> 0534677 (microservices testing local)
